// core/models/allergens.ts
export type Allergen =
	| 'GLUTEN_WHEAT'
	| 'CRUSTACEANS'
	| 'EGGS'
	| 'FISH'
	| 'PEANUTS'
	| 'SOYA'
	| 'MILK_LACTOSE'
	| 'NUTS'
	| 'CELERY'
	| 'MUSTARD'
	| 'SESAME'
	| 'SULPHUR_DIOXIDE'
	| 'LUPIN'
	| 'MOLLUSCS';

const allergens: Allergen[] = [
	'GLUTEN_WHEAT',
	'CRUSTACEANS',
	'EGGS',
	'FISH',
	'PEANUTS',
	'SOYA',
	'MILK_LACTOSE',
	'NUTS',
	'CELERY',
	'MUSTARD',
	'SESAME',
	'SULPHUR_DIOXIDE',
	'LUPIN',
	'MOLLUSCS',
];

export default allergens;